курсовая базы данных/requirements.txt
bottle==0.13.0
pg8000==1.30.3
python-dotenv==1.0.0
Faker==19.13.0
python-dateutil==2.8.2 